refactor: rename pynumcalc.* modules
