+++ pynumcalc/_typedef.py
